+++ src/MrCatzEvent.php
@@ -24,6 +24,7 @@ final class MrCatzEvent
     public const PREPARE_DELETE = 'prepareDeleteData';
     public const BULK_DELETE = 'bulkDeleteData';
     public const BULK_ACTION = 'bulkActionData';
+    public const BULK_ACTION_OPEN = 'bulkActionOpen';
     public const REFRESH_TABLE = 'refreshDataTable';
     public const INLINE_UPDATE = 'inlineUpdateData';
     public const ROW_CLICK = 'rowClickData';